migrate: new GltManager contains glTF load+save logic. Migrated from PreCacheSceneManager.
