+++ Assets/UnZENity-Core/Scripts/GameGlobals.cs
@@ -13,6 +13,7 @@ namespace GUZ.Core
         public static GameConfiguration Config => Instance.Config;
         public static GameSettings Settings => Instance.Settings;
         public static LoadingManager Loading => Instance.Loading;
+        public static GltManager Glt => Instance.Glt;
         public static PlayerManager Player => Instance.Player;
         public static SkyManager Sky => Instance.Sky;
         public static GameTime Time => Instance.Time;
