+++ Assets/UnZENity-Lab/Scripts/LabBootstrapper.cs
@@ -43,6 +43,7 @@ namespace GUZ.Lab
 
         public GameSettings Settings => _settings;
         public LoadingManager Loading => null;
+        public GltManager Glt => null;
         public PlayerManager Player => null;
         public SkyManager Sky => null;
         public GameTime Time => null;
